001_django_throttling/requirements.txt
Django==4.0.7
djangorestframework==3.13.1
gunicorn==20.1.0
redis==4.3.4
